feat: Add one-on-one pairing generation command
Add functionality to generate one-on-one pairings for participants.
Introduce a new command `one_on_ones` that takes a list of names, number of organizers, number of rounds, and number of tries to generate pairings. The organizers are prioritized in the pairing process, and participants are paired in a round-robin fashion.