docs(meta): Add README placeholder for Google scripts
Add a placeholder README file in the `meta` directory for future documentation on using scripts to automate tasks with Google Drive/Docs. The content is pending and will be written by Diego.


docs(meta): Add README placeholder for Google scripts

Add a placeholder README file in the `meta` directory for future documentation on using scripts to automate tasks with Google Drive/Docs. The content is pending and will be written by Diego.